Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/dom/events/shadow-relatedTarget-expected.txt



FAIL relatedTarget should not leak at capturing phase, at window object. assert_equals: expected Element node <div id="host"></div> but got Element node <input id="shadowInput"></input>
PASS relatedTarget should not leak at target.
